app.get('/api/candidates',(req,res)=>{
const sql= ;
const params=[];
db.all(sql,params, (err, rows) => {
    if(err){
        res.status(500).json({error:err.message});
    }
  res.json({
      message:'success',
      data: rows
  });
});
});
We'll use the get() method to retrieve all the candidates from the candidates table. This route is designated with the endpoint /api/candidates. Remember, the api in the URL signifies that this is an API endpoint.

The callback function will handle the client's request and the database's response. The request-response pattern should look familiar from the last module.

We'll wrap the get() method around the database call we made earlier and modify it a bit. The SQL statement SELECT * FROM candidates is assigned to the sql variable. Here we set the params assignment to an empty array because there are no placeholders in the SQL statement.

We use the all() method from the database object to retrieve all the rows. In the error-handling conditional, instead of logging the error, we'll send a status message 500 and the error message in a JSON object. The 500 status code means there was a server error—different than a 404, which indicates a user request error. The return statement will exit out of the database call once an error is encountered.

If there was no error, then err is null and the response is sent back using the following statement:
es.json({
  message: 'success',
  data: rows
});
Again, instead of logging the result, rows, from the database, we'll send this response as a JSON object to the browser, using res in the Express.js route callback.